Add files via upload
Mock User Guide: This is an excerpt from a User Guide to Interdimensional Portal Travel. It was inspired by the  Meow Wolf art collective’s interactive art installation “Omega Mart.” I used my photos of the exhibit, the Meow Wolf fan wiki, and my imagination to create this mock user guide. The “signs of portal sickness” are from a poster in Omega Mart. I borrowed the name Nimsesku from an interdimensional hamster in a different Meow Wolf exhibit. Everything else is my own work. The audience for this user guide is interdimensional travelers who are familiar with forkles and nimseskus.